split files, rename and cleanup
